+++ src/index.ts
@@ -1,1 +1,4 @@
 /// <reference path="../index.d.ts" />
+
+export * from './interfaces'
+export * from './util/ordered'