+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## DEVELOPMENT
-### 0.1.0
+### 0.2.0
+`+` Import a bunch more core util classes and interfaces from `vx-util`  
+`*` Refactor `Functional` and `FunctionalAsync`  
 
-### 0.0.0
+### 0.1.0
 `*` Import implementation from `vx-util`